requirements.txt
alembic==1.17.2
altgraph==0.17.4
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pycparser==2.23
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
waitress==3.0.2
Werkzeug==3.1.4
